+++ protocol/rtmp/src/remuxer/rtsp2rtmp.rs
@@ -31,7 +31,6 @@ pub struct Rtsp2RtmpRemuxerSession {
     app_name: String,
     stream_name: String,
 
-    publishe_id: Uuid,
     //RTSP
     data_receiver: FrameDataReceiver,
     stream_path: String,
@@ -73,7 +72,7 @@ impl Rtsp2RtmpRemuxerSession {
             event_producer: event_producer.clone(),
 
             subscribe_id: Uuid::new(RandomDigitCount::Four),
-            publishe_id: Uuid::new(RandomDigitCount::Four),
+
             video_clock_rate: 1000,
             audio_clock_rate: 1000,
             base_audio_timestamp: 0,
@@ -93,30 +92,20 @@ impl Rtsp2RtmpRemuxerSession {
 
     pub async fn publish_rtmp(&mut self) -> Result<(), RtmpRemuxerError> {
         self.rtmp_handler
-            .publish_to_channels(
-                self.app_name.clone(),
-                self.stream_name.clone(),
-                self.publishe_id,
-                0,
-            )
+            .publish_to_channels(self.app_name.clone(), self.stream_name.clone(), 0)
             .await?;
         Ok(())
     }
 
     pub async fn unpublish_rtmp(&mut self) -> Result<(), RtmpRemuxerError> {
         self.rtmp_handler
-            .unpublish_to_channels(
-                self.app_name.clone(),
-                self.stream_name.clone(),
-                self.publishe_id,
-            )
+            .unpublish_to_channels(self.app_name.clone(), self.stream_name.clone())
             .await?;
         Ok(())
     }
 
     pub async fn subscribe_rtsp(&mut self) -> Result<(), RtmpRemuxerError> {
         let (event_result_sender, event_result_receiver) = oneshot::channel();
-
         let sub_info = SubscriberInfo {
             id: self.subscribe_id,
             sub_type: SubscribeType::PlayerRtmp,
@@ -141,7 +130,7 @@ impl Rtsp2RtmpRemuxerSession {
             });
         }
 
-        let receiver = event_result_receiver.await??;
+        let receiver = event_result_receiver.await??.0;
         self.data_receiver = receiver.frame_receiver.unwrap();
         Ok(())
     }